Rewrite Consul bootstrapping and networking
Instead of relying on a single interface name, which may not even be
accurate in a Swarm when using multiple networks as containers are
connected to the networks with seemingly random interface names, we'll
instead forward through the service name.

Using this service name, we can query Docker Swarm's inherent DNS
entries to determine the correct IP addresses to bind Consul to.

This does mean that the environment variables defining Consul have to be
correctly defined, but sensible defaults have been used.