Allow MySQL replica to take leadership if leader is disconnected
In the case where a leader node went down, a new leader was unable to be elected from the replicas because it failed a check on the IO state which was expecting the master to send an event. We now will also allow a leadership spill if the IO state says the leader cannot be reached but Consul can be reached, as the other nodes will detect the new leader through Consul.

We still rely on the replica's SQL state to have read all the relay log before it can be elected, so theoretically this replica *should* be up to date.