Use Consul to determine if a snapshot is pending
If a node goes down while it is taking a snapshot, a pending snapshot may gridlock some functionality that waits for pending snapshots to be completed, because the pending snapshot will be orphaned.

Since Consul now tracks snapshotting state as well, we'll confirm with Consul if a snapshot is taking place - if no nodes are snapshotting, we'll consider the pending snapshot orphaned and remove it in order for everything to continue.