Add ability to prime a snapshot using an SQL dump
To allow people to bootstrap larger datasets, the MCM now accepts an
"execute_file" command to run an SQL file against the application DB,
and then snapshot the final database state. This way, a cluster could be
bootstrapped with its initial data in place and then brought online from
the snapshots.